refactor(docs): extract code group sync into utils helper
